pyproject.toml
[project]
name = "nonebot-plugin-weiweibot"
version = "0.2.1"
description = "a simple weiwei bot"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Alesso", email = "intro@j4cka.cc" }]
dependencies = [
    "nonebot2>=2.2.0",
    "nonebot-adapter-onebot>=2.1.3",
    "python-gnupg>=0.5.4",
    "algoliasearch>=4.22.1",
    "filetype>=1.2.0",
    "httpx>=0.28.1",
]
license = { text = "MIT" }
keywords = ["nonebot"]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = ["nonebot_plugin_localstore"]
plugin_dirs = ["plug"]
builtin_plugins = []

[tool.pdm]
distribution = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.urls]
homepage = "https://github.com/SwedishDoveCooker/nonebot-plugin-weiweibot"
repository = "https://github.com/SwedishDoveCooker/nonebot-plugin-weiweibot"

[tool.pdm.build]
includes = ["plug/nonebot-plugin-weiweibot"]
